backend/requirements.txt
fastapi
uvicorn[standard]
oracledb
python-dotenv
pydantic
sqlalchemy
passlib[bcrypt]
python-multipart
python-jose
